converted main printer rows to component (still needs work)
